[CI/Build] [1/3] Reorganize entrypoints tests (#5526)
